cab-project/backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
geoalchemy2==0.14.2

# External APIs
googlemaps==4.10.0
stripe==7.8.0

# Real-time and Background Tasks
python-socketio==5.10.0
celery==5.3.4
redis==5.0.1

# PDF Generation
WeasyPrint==60.2
wkhtmltopdf==0.2
jinja2==3.1.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP and Utilities
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1